Notification component connected to Redux, can show oldData msg (WIP)
