+++ src/components/notification/index.js
@@ -1,74 +1,34 @@
-// Docs https://material-ui.com/demos/snackbars/
-import React from 'react'
-import PropTypes from 'prop-types'
-import { withStyles } from '@material-ui/core/styles'
-import Snackbar from '@material-ui/core/Snackbar'
-import IconButton from '@material-ui/core/IconButton'
-import CloseIcon from '@material-ui/icons/Close'
+import { connect } from 'react-redux'
+import { get } from 'lodash'
 
-const styles = theme => ({
-  close: {
-    width: theme.spacing.unit * 4,
-    height: theme.spacing.unit * 4,
-  },
+import Notification from './component'
+import { shouldUpdate } from '../../utils/should-update'
+
+const mapStateToProps = state => ({
+  lists: get(state, 'lists', {}),
 })
 
-class Notification extends React.Component {
-  constructor(props) {
-    super(props)
-    this.state = {
-      open: typeof props.message !== 'undefined' || false,
-    }
-  }
+const mapDispatchToProps = dispatch => ({ dispatch })
 
-  handleClose = (event, reason) => {
-    if (reason === 'clickaway') {
-      return
-    }
+const mergeProps = (stateProps, dispatchProps, ownProps) => {
+  const shouldAppUpdate = shouldUpdate(get(stateProps, 'lists.ts'))
 
-    this.setState({ open: false })
-  }
-
-  render() {
-    const { classes, message } = this.props
-    return (
-      <div>
-        <Snackbar
-          anchorOrigin={{
-            vertical: 'bottom',
-            horizontal: 'right',
-          }}
-          open={this.state.open}
-          autoHideDuration={10000}
-          onClose={this.handleClose}
-          ContentProps={{
-            'aria-describedby': 'message-id',
-          }}
-          message={<span id="message-id">{message}</span>}
-          action={[
-            <IconButton
-              key="close"
-              aria-label="Close"
-              color="inherit"
-              className={classes.close}
-              onClick={this.handleClose}
-            >
-              <CloseIcon />
-            </IconButton>,
-          ]}
-        />
-      </div>
-    )
+  // TODO:
+  // - we need to add a periodic check to make this appear, otherwise the components
+  // will refresh the data anyway will mounting (see their componentDidMount method)
+  return {
+    ...ownProps,
+    message: shouldAppUpdate
+      ? 'You are working with old data. Please reload the page.'
+      : ownProps.message,
+    action: shouldUpdate
+      ? {
+          onClick: () => window.location.reload(true),
+          text: 'Reload',
+        }
+      : ownProps.action,
   }
 }
 
-Notification.propTypes = {
-  classes: PropTypes.object.isRequired, // eslint-disable-line
-  message: PropTypes.string,
-}
-
-Notification.defaultProps = {
-  message: undefined,
-}
-
-export default withStyles(styles)(Notification)
+const NotificationContainer = connect(mapStateToProps, mapDispatchToProps, mergeProps)(Notification)
+export default NotificationContainer
